requirements.txt
databricks-sdk==0.50.0 
databricks-langchain>=0.4.0 
databricks-agents 
langchain==0.3.25 
langchain_core==0.3.59 
mlflow[databricks] 
databricks-vectorsearch==0.55 
openai==1.68.2
openai-agents==0.0.7
pydantic==2.10.1
rich==14.0.0
unitycatalog-ai==0.3.0 
unitycatalog-client
streamlit==1.44.1
streamlit-avatar==0.1.3
